feat: navigation item docs
